+++ src/AElf.Contracts.TestKit/ContractTestModule.cs
@@ -110,10 +110,12 @@ namespace AElf.Contracts.TestKit
 
             context.Services.AddTransient<IAccount, Account>();
             context.Services.AddTransient<IContractTesterFactory, ContractTesterFactory>();
-            context.Services.AddTransient<ITransactionExecutor, TransactionExecutor>();
+            context.Services.AddTransient<ITestTransactionExecutor, TestTransactionExecutor>();
             context.Services.AddSingleton<IBlockTimeProvider, BlockTimeProvider>();
             context.Services.AddSingleton<ITxHub, MockTxHub>();
-            context.Services.Replace(ServiceDescriptor.Singleton<ILocalParallelTransactionExecutingService, LocalTransactionExecutingService>());
+            context.Services.Replace(ServiceDescriptor
+                .Singleton<ITransactionExecutingService,
+                    Kernel.SmartContract.Application.LocalTransactionExecutingService>());
             context.Services.AddSingleton<IChargeFeeStrategy, ZeroContractChargeFeeStrategy>();
         }
