+++ src/AElf.Kernel.SmartContract.Parallel/Application/BlockParallelExecutingService.cs
@@ -1,0 +1,32 @@
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using AElf.Kernel.SmartContract.Application;
+using AElf.Kernel.SmartContract.Domain;
+using AElf.Kernel.SmartContractExecution.Application;
+using AElf.Types;
+
+//TODO: namespace
+namespace AElf.Kernel.SmartContract.Parallel
+{
+    public class BlockParallelExecutingService: BlockExecutingService
+    {
+        public BlockParallelExecutingService(ITransactionExecutingService transactionExecutingService,
+            IBlockchainStateService blockchainStateService) : base(transactionExecutingService, blockchainStateService)
+        {
+        }
+
+        protected override async Task<Block> FillBlockAfterExecutionAsync(BlockHeader blockHeader, List<Transaction> transactions, ReturnSetCollection returnSetCollection)
+        {
+            var block = await base.FillBlockAfterExecutionAsync(blockHeader, transactions, returnSetCollection);
+            if(returnSetCollection.Conflict.Count > 0)
+            {
+                await EventBus.PublishAsync(new ConflictingTransactionsFoundInParallelGroupsEvent(
+                    block.Header, returnSetCollection.Executed.Concat(returnSetCollection.Unexecutable).ToList(),
+                    returnSetCollection.Conflict
+                ));
+            }
+            return block;
+        }
+    }
+}
