+++ src/AElf.Kernel.SmartContract.Parallel/ParallelExecutionModule.cs
@@ -1,3 +1,4 @@
+using AElf.Kernel.SmartContract.Application;
 using AElf.Kernel.SmartContractExecution.Application;
 using AElf.Modularity;
 using Microsoft.Extensions.DependencyInjection;
@@ -10,10 +11,9 @@ namespace AElf.Kernel.SmartContract.Parallel
     {
         public override void ConfigureServices(ServiceConfigurationContext context)
         {
-            //
             context.Services.AddStoreKeyPrefixProvide<ContractRemarks>("cr");
-
-            context.Services.AddTransient<IBlockExecutingService, BlockExecutingWithParallelService>();
+            context.Services.AddTransient<IBlockExecutingService, BlockParallelExecutingService>();
+            context.Services.AddSingleton<ITransactionExecutingService, LocalParallelTransactionExecutingService>();
         }
     }
 }
