+++ src/AElf.Kernel.SmartContract/Application/ITransactionExecutor.cs
@@ -4,13 +4,14 @@ using System.Threading.Tasks;
 
 namespace AElf.Kernel.SmartContract.Application
 {
-    //TODO: the inherit is very very strange
-    public interface ILocalTransactionExecutingService: ILocalParallelTransactionExecutingService
-    {
-    }
-    public interface ILocalParallelTransactionExecutingService
+    public interface ITransactionExecutor
     {
         Task<List<ExecutionReturnSet>> ExecuteAsync(TransactionExecutingDto transactionExecutingDto,
             CancellationToken cancellationToken, bool throwException = false);
+    }
+
+    public interface ITransactionExecutingService : ITransactionExecutor
+    {
+
     }
 }
