+++ src/AElf.Kernel.SmartContractExecution/Application/BlockExecutingService.cs
@@ -17,16 +17,15 @@ namespace AElf.Kernel.SmartContractExecution.Application
 {
     public class BlockExecutingService : IBlockExecutingService, ITransientDependency
     {
-        //TODO: should know nothing about Parallel
-        private readonly ILocalParallelTransactionExecutingService _executingService;
+        private readonly ITransactionExecutingService _transactionExecutingService;
         private readonly IBlockchainStateService _blockchainStateService;
         public ILocalEventBus EventBus { get; set; }
         public ILogger<BlockExecutingService> Logger { get; set; }
 
-        public BlockExecutingService(ILocalParallelTransactionExecutingService executingService,
+        public BlockExecutingService(ITransactionExecutingService transactionExecutingService,
             IBlockchainStateService blockchainStateService)
         {
-            _executingService = executingService;
+            _transactionExecutingService = transactionExecutingService;
             _blockchainStateService = blockchainStateService;
             EventBus = NullLocalEventBus.Instance;
         }
@@ -46,7 +45,7 @@ namespace AElf.Kernel.SmartContractExecution.Application
             var nonCancellable = nonCancellableTransactions.ToList();
             var cancellable = cancellableTransactions.ToList();
             var nonCancellableReturnSets =
-                await _executingService.ExecuteAsync(
+                await _transactionExecutingService.ExecuteAsync(
                     new TransactionExecutingDto {BlockHeader = blockHeader, Transactions = nonCancellable},
                     CancellationToken.None, true);
             Logger.LogTrace("Executed non-cancellable txs");
@@ -56,7 +55,7 @@ namespace AElf.Kernel.SmartContractExecution.Application
 
             if (!cancellationToken.IsCancellationRequested && cancellable.Count > 0)
             {
-                cancellableReturnSets = await _executingService.ExecuteAsync(
+                cancellableReturnSets = await _transactionExecutingService.ExecuteAsync(
                     new TransactionExecutingDto
                     {
                         BlockHeader = blockHeader,
