+++ test/AElf.Contracts.Economic.TestBase/EconomicContractsTestModule.cs
@@ -22,7 +22,7 @@ namespace AElf.Contracts.Economic.TestBase
         {
             Configure<ContractOptions>(o => o.ContractDeploymentAuthorityRequired = false );
 
-            context.Services.AddSingleton<ITransactionExecutor, EconomicTransactionExecutor>();
+            context.Services.AddSingleton<ITestTransactionExecutor, EconomicTestTransactionExecutor>();
             context.Services.AddSingleton<ITriggerInformationProvider, AEDPoSTriggerInformationProvider>();
             context.Services.AddSingleton<IBlockValidationService, MockBlockValidationService>();
             context.Services.AddSingleton<IPreExecutionPlugin, FeeChargePreExecutionPlugin>();
