+++ test/AElf.Kernel.SmartContractExecution.Tests/SmartContractExecutionTestAElfModule.cs
@@ -15,7 +15,7 @@ namespace AElf.Kernel.SmartContractExecution
     {
         public override void ConfigureServices(ServiceConfigurationContext context)
         {
-            context.Services.Replace(ServiceDescriptor.Singleton<ILocalParallelTransactionExecutingService, LocalTransactionExecutingService>());
+            context.Services.Replace(ServiceDescriptor.Singleton<ILocalParallelTransactionExecutingService, TransactionExecutor>());
             context.Services.AddTransient(provider =>
             {
                 var mockService = new Mock<IDeployedContractAddressService>();
