+++ test/AElf.WebApp.Application.TestBase/WebAppTestAElfModule.cs
@@ -92,7 +92,7 @@ namespace AElf.WebApp.Application
                 .AddTransient<ITransactionValidationProvider, TransactionFromAddressBalanceValidationProvider>();
             context.Services.AddTransient<ITransactionValidationProvider, TransactionToAddressValidationProvider>();
             context.Services.AddSingleton<IPreExecutionPlugin, FeeChargePreExecutionPlugin>();
-            context.Services.Replace(ServiceDescriptor.Singleton<ILocalParallelTransactionExecutingService, LocalTransactionExecutingService>());
+            context.Services.Replace(ServiceDescriptor.Singleton<ILocalParallelTransactionExecutingService, TransactionExecutor>());
         }
     }
 }
